# controller/user.go
package controller

import (
	"context"
	"errors"
	"fmt"
	"net/http"

	"github.com/alert666/api-server/base/bind"
	"github.com/alert666/api-server/base/constant"
	v1 "github.com/alert666/api-server/service/v1"
	"github.com/gin-contrib/sessions"
	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
)

type UserController interface {
	UserLoginController(c *gin.Context)
	UserLogoutController(c *gin.Context)
	UserCreateController(c *gin.Context)
	UserUpdateByAdminController(c *gin.Context)
	UserUpdateBySelfController(c *gin.Context)
	UserDeleteController(c *gin.Context)
	UserQueryController(c *gin.Context)
	UserListController(c *gin.Context)
	UserInfoController(c *gin.Context)
	OAuth2LoginController(c *gin.Context)
	OAuth2CallbackController(c *gin.Context)
	OAuth2ProviderController(c *gin.Context)
	OAuth2ActivateController(c *gin.Context)
	GetUserOption(c *gin.Context)
}

type UserControllerImpl struct {
	userServicer v1.UserServicer
}

func NewUserController(userServicer v1.UserServicer) UserController {
	return &UserControllerImpl{
		userServicer: userServicer,
	}
}

// UserLoginController 用户登录
// @Summary 用户登录
// @Description 使用邮箱和密码登录，返回用户信息和 Token
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param data body types.UserLoginRequest true "登录请求参数"
// @Success 200 {object} types.Response{} "登录成功"
// @Router /api/v1/user/login [post]
func (receiver *UserControllerImpl) UserLoginController(c *gin.Context) {
	bind.ResponseWithData(c, receiver.userServicer.Login, bind.BindTypeJson)
}

// UserLogoutController 用户注销
// @Summary 用户注销
// @Description 用户注销，清空 Token
// @Tags 用户管理
// @Accept json
// @Produce json
// @Success 200 {object} types.Response "注销成功"
// @Router /api/v1/user/logout [post]
func (receiver *UserControllerImpl) UserLogoutController(c *gin.Context) {
	bind.ResponseNoBind(c, receiver.userServicer.Logout)
}

// UserCreateController 用户创建
// @Summary 用户创建
// @Description 创建用户同时可以设置角色
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param data body types.UserCreateRequest true "创建请求参数"
// @Success 200 {object} types.Response "创建成功"
// @Router /api/v1/user/register [post]
func (receiver *UserControllerImpl) UserCreateController(c *gin.Context) {
	bind.ResponseOnlySuccess(c, receiver.userServicer.CreateUser, bind.BindTypeJson)
}

// UserUpdateByAdminController 用户更新
// @Summary 用户更新
// @Description 更新用户信息，可以更新角色
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param data body types.UserUpdateAdminRequest true "更新请求参数"
// @Success 200 {object} types.Response "更新成功"
// @Router /api/v1/user/:id [put]
func (receiver *UserControllerImpl) UserUpdateByAdminController(c *gin.Context) {
	bind.ResponseOnlySuccess(c, receiver.userServicer.UpdateUserByAdmin, bind.BindTypeUri, bind.BindTypeJson)
}

// UserUpdateBySelfController 用户更新自己的信息
// @Summary 用户更新自己的信息
// @Description 更新用户信息，不能更新角色
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param data body types.UserUpdateSelfRequest true "更新请求参数"
// @Success 200 {object} types.Response "更新成功"
// @Router /api/v1/user/self [put]
func (receiver *UserControllerImpl) UserUpdateBySelfController(c *gin.Context) {
	bind.ResponseOnlySuccess(c, receiver.userServicer.UpdateUserBySelf, bind.BindTypeJson)
}

// UserDeleteController 用户删除
// @Summary 用户删除
// @Description 删除用户，只能管理员删除
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param data body types.IDRequest true "删除请求参数"
// @Success 200 {object} types.Response "删除成功"
// @Router /api/v1/user/:id [delete]
func (receiver *UserControllerImpl) UserDeleteController(c *gin.Context) {
	bind.ResponseOnlySuccess(c, receiver.userServicer.DeleteUser, bind.BindTypeUri)
}

// UserQueryController 用户查询
// @Summary 用户查询
// @Description 使用 id 查询用户的信息和用户的角色
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param data body types.IDRequest true "查询请求参数"
// @Success 200 {object} types.Response{} "查询成功"
// @Router /api/v1/user/:id [get]
func (receiver *UserControllerImpl) UserQueryController(c *gin.Context) {
	bind.ResponseWithData(c, receiver.userServicer.QueryUser, bind.BindTypeUri)
}

// UserInfoController 用户获取自己的信息
// @Summary 用户获取自己的信息
// @Description 使用 id 查询用户的信息和用户的角色
// @Tags 用户管理
// @Accept json
// @Produce json
// @Success 200 {object} types.Response{} "查询成功"
// @Router /api/v1/user/info [get]
func (receiver *UserControllerImpl) UserInfoController(c *gin.Context) {
	bind.ResponseWithDataNoBind(c, receiver.userServicer.Info)
}

// UserListController 用户列表
// @Summary 用户列表
// @Description 使用分页查询用户的信息, 支持根据 name, email, mobile, department 查询
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param data query types.UserListRequest true "查询请求参数"
// @Success 200 {object} types.Response{} "登录成功"
// @Router /api/v1/user [get]
func (receiver *UserControllerImpl) UserListController(c *gin.Context) {
	bind.ResponseWithData(c, receiver.userServicer.ListUser, bind.BindTypeQuery)
}

// OAuth2LoginController OAuth 登录
// @Summary OAuth 登录
// @Description 使用 OAuth 登录，返回用户信息和 Token
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param provider query string false "OAuth2 provider 名称（如 feishu、keycloak）"
// @Success 302 {string} string "重定向到 OAuth 登录页面"
// @Router /api/v1/oauth2/login [get]
func (receiver *UserControllerImpl) OAuth2LoginController(c *gin.Context) {
	session := sessions.Default(c)
	state := uuid.New().String()
	session.Set("state", state)
	provider := c.Query("provider")
	if provider != "" {
		session.Set("provider", provider)
	}

	if err := session.Save(); err != nil {
		bind.ResponseError(nil, c, fmt.Errorf("save session failed: %w", err))
		return
	}
	url, err := receiver.userServicer.OAuth2Login(provider, state)
	if err != nil {
		bind.ResponseError(nil, c, err)
		return
	}
	c.Redirect(http.StatusFound, url)
}

// OAuth2CallbackController OAuth 回调
// @Summary OAuth 回调
// @Description 使用 OAuth 回调，返回用户信息和 Token
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param data query types.OAuthLoginRequest true "回调请求参数"
// @Success 200 {object} types.Response{} "登录成功"
// @Router /api/v1/oauth2/callback [get]
func (receiver *UserControllerImpl) OAuth2CallbackController(c *gin.Context) {
	session := sessions.Default(c)
	stateSession := session.Get("state")
	providerSession := session.Get("provider")
	state := c.Query("state")
	if state == "" {
		bind.ResponseError(nil, c, errors.New("state is empty"))
		return
	}
	if state != stateSession {
		bind.ResponseError(nil, c, errors.New("state invalid"))
		return
	}
	var providerStr string
	if providerSession != nil {
		if s, ok := providerSession.(string); ok {
			providerStr = s
		}
	}
	if providerStr == "" {
		bind.ResponseError(nil, c, errors.New("provider is empty"))
		return
	}
	ctx := context.WithValue(c.Request.Context(), constant.ProviderContextKey, providerStr)
	c.Request = c.Request.WithContext(ctx)
	bind.ResponseWithData(c, receiver.userServicer.OAuth2Callback, bind.BindTypeQuery)
}

// OAuth2ProviderController OAuth2 提供商列表
// @Summary OAuth2 提供商列表
// @Description 获取 OAuth2 提供商列表
// @Tags 用户管理
// @Accept json
// @Produce json
// @Success 200 {object} types.Response{} "获取成功"
// @Router /api/v1/oauth2/provider [get]
func (receiver *UserControllerImpl) OAuth2ProviderController(c *gin.Context) {
	bind.ResponseWithDataNoBind(c, receiver.userServicer.OAuth2Provider)
}

// OAuth2ActivateController OAuth2 激活
// @Summary OAuth2 激活
// @Description 使用 OAuth2 激活，返回用户信息和 Token
// @Tags 用户管理
// @Accept json
// @Produce json
// @Param data body types.OAuthActivateRequest true "激活请求参数"
// @Success 200 {object} types.Response{} "激活成功"
// @Router /api/v1/oauth2/:id [post]
func (receiver *UserControllerImpl) OAuth2ActivateController(c *gin.Context) {
	session := sessions.Default(c)
	stateSession := session.Get("state")
	state := c.Query("state")
	if state == "" {
		bind.ResponseError(nil, c, errors.New("state is empty"))
		return
	}
	if state != stateSession {
		bind.ResponseError(nil, c, errors.New("state invalid"))
		return
	}
	bind.ResponseWithData(c, receiver.userServicer.OAuth2Activate, bind.BindTypeUri, bind.BindTypeJson)
}

// GetUserOption 获取用户 Options
// @Summary 获取用户 Options
// @Description 获取用户 Options
// @Tags 用户管理
// @Accept json
// @Produce json
// @Success 200 {object} types.Response{} "查询成功"
// @Router /api/v1/user/options [get]
func (receiver *UserControllerImpl) GetUserOption(c *gin.Context) {
	bind.ResponseWithDataNoBind(c, receiver.userServicer.GetUserOptions)
}